feat(refac): Refactoring packages
