+++ src/main/java/com/example/desafioquality/dto/PropertyTotalAreaDto.java
@@ -1,4 +1,4 @@
-package com.example.desafioquality.controller.dto;
+package com.example.desafioquality.dto;
 
 
 import lombok.Getter;
